+++ src/main/scala/io/pg/config/format.scala
@@ -1,24 +1,48 @@
 package io.pg.config
 
-sealed trait TextMatcher extends Product with Serializable
+import cats.data.NonEmptyList
+import io.circe.Codec
+import io.circe.generic.extras.Configuration
+import io.circe.generic.extras.ConfiguredJsonCodec
 
-object TextMatcher {
-  final case class Equals(value: String) extends TextMatcher
-  final case class Matches(pattern: String) extends TextMatcher
+object circe {
+  implicit val circeConfig: Configuration =
+    Configuration.default.withDiscriminator("kind").withSnakeCaseConstructorNames
 }
 
-sealed trait Matcher extends Product with Serializable {
-  def and(another: Matcher): Matcher = Matcher.Many(List(this, another))
-  def or(another: Matcher): Matcher = Matcher.OneOf(List(this, another))
+import circe.circeConfig
+
+@ConfiguredJsonCodec()
+sealed trait Result extends Product with Serializable
+
+object Result {
+  case object Ok extends Result
+
+  final case class NotOk(mismatches: NonEmptyList[Mismatch]) extends Result
 }
 
-object Matcher {
-  final case class Author(email: TextMatcher) extends Matcher
-  final case class Description(text: TextMatcher) extends Matcher
-  final case class PipelineStatus(status: String) extends Matcher
-  final case class Many(values: List[Matcher]) extends Matcher
-  final case class OneOf(values: List[Matcher]) extends Matcher
-  final case class Not(underlying: Matcher) extends Matcher
+@ConfiguredJsonCodec()
+sealed trait Mismatch extends Product with Serializable
 
-  val always: Matcher = Many(Nil)
+object Mismatch {
+  final case class Status(expected: io.pg.config.Status, actual: io.pg.config.Status) extends Mismatch
+  final case class Author(expected: TextRule, actual: String) extends Mismatch
+  final case class Description(expected: TextRule, actual: String) extends Mismatch
+  final case class NoneMatched(mismatches: NonEmptyList[Mismatch]) extends Mismatch
+}
+
+@ConfiguredJsonCodec()
+sealed trait TextRule extends Product with Serializable
+
+object TextRule {
+  final case class Equal(expected: String) extends TextRule
+  final case class Matches(pattern: String) extends TextRule
+
+}
+
+final case class Status(value: String) extends AnyVal
+
+object Status {
+  implicit val codec: Codec[Status] = io.circe.generic.extras.semiauto.deriveUnwrappedCodec
+
 }
